pyproject.toml
[project]
name = "agency"
version = "4.0.0"
description = "AI marketing agency. Agent-first, HITL optional."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0,<3.0",
    "typer>=0.9.0",
    "rich>=13.0",
    "httpx>=0.25",
    "google-genai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.1.0",
]

[project.scripts]
agency = "agency.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
